Split toml/test.ts into separate files
Handling issue #4593. Took tests from toml/test.ts and split into toml/parse_tests.ts and toml/stringify_tests.ts
